Converted parts of the integration test for ReportListSettings to unit tests
