+++ tests/integration/Frontend/ReportList/ReportListSettingsTest.php
@@ -2,44 +2,19 @@
 namespace abrain\Einsatzverwaltung\Frontend\ReportList;
 
 use WP_UnitTestCase;
-use abrain\Einsatzverwaltung\Frontend\ReportList\Settings as ReportListSettings;
 
+/**
+ * Class ReportListSettingsTest
+ * @package abrain\Einsatzverwaltung\Frontend\ReportList
+ */
 class ReportListSettingsTest extends WP_UnitTestCase
 {
     const DEFAULT_NTHCHILD = 'even';
     const DEFAULT_ZEBRACOLOR = '#eee';
 
-    public function testGetZebraNthChildArg()
-    {
-        $reportListSettings = new ReportListSettings();
-        delete_option('einsatzvw_list_zebra_nth');
-        $this->assertEquals(self::DEFAULT_NTHCHILD, $reportListSettings->getZebraNthChildArg());
-        update_option('einsatzvw_list_zebra_nth', '');
-        $this->assertEquals(self::DEFAULT_NTHCHILD, $reportListSettings->getZebraNthChildArg());
-        update_option('einsatzvw_list_zebra_nth', 'invalid');
-        $this->assertEquals(self::DEFAULT_NTHCHILD, $reportListSettings->getZebraNthChildArg());
-        update_option('einsatzvw_list_zebra_nth', 5);
-        $this->assertEquals(self::DEFAULT_NTHCHILD, $reportListSettings->getZebraNthChildArg());
-        update_option('einsatzvw_list_zebra_nth', 'odd');
-        $this->assertEquals('odd', $reportListSettings->getZebraNthChildArg());
-        update_option('einsatzvw_list_zebra_nth', 'even');
-        $this->assertEquals('even', $reportListSettings->getZebraNthChildArg());
-    }
-
-    public function testSanitizeZebraNthChildArg()
-    {
-        $reportListSettings = new ReportListSettings();
-        $this->assertEquals('odd', $reportListSettings->sanitizeZebraNthChildArg('odd'));
-        $this->assertEquals('even', $reportListSettings->sanitizeZebraNthChildArg('even'));
-        $this->assertEquals(self::DEFAULT_NTHCHILD, $reportListSettings->sanitizeZebraNthChildArg(''));
-        $this->assertEquals(self::DEFAULT_NTHCHILD, $reportListSettings->sanitizeZebraNthChildArg('invalid'));
-        $this->assertEquals(self::DEFAULT_NTHCHILD, $reportListSettings->sanitizeZebraNthChildArg(2));
-        $this->assertEquals(self::DEFAULT_NTHCHILD, $reportListSettings->sanitizeZebraNthChildArg(null));
-    }
-
     public function testGetZebraColor()
     {
-        $reportListSettings = new ReportListSettings();
+        $reportListSettings = new Settings();
         delete_option('einsatzvw_list_zebracolor');
         $this->assertEquals(self::DEFAULT_ZEBRACOLOR, $reportListSettings->getZebraColor());
         update_option('einsatzvw_list_zebracolor', '');
@@ -54,20 +29,5 @@ class ReportListSettingsTest extends WP_UnitTestCase
         $this->assertEquals(self::DEFAULT_ZEBRACOLOR, $reportListSettings->getZebraColor());
         update_option('einsatzvw_list_zebracolor', 3);
         $this->assertEquals(self::DEFAULT_ZEBRACOLOR, $reportListSettings->getZebraColor());
-    }
-
-    public function testIsZebraTable()
-    {
-        $reportListSettings = new ReportListSettings();
-        delete_option('einsatzvw_list_zebra');
-        $this->assertTrue($reportListSettings->isZebraTable());
-        update_option('einsatzvw_list_zebra', '0');
-        $this->assertFalse($reportListSettings->isZebraTable());
-        update_option('einsatzvw_list_zebra', '1');
-        $this->assertTrue($reportListSettings->isZebraTable());
-        update_option('einsatzvw_list_zebra', 'invalid');
-        $this->assertTrue($reportListSettings->isZebraTable());
-        update_option('einsatzvw_list_zebra', 0);
-        $this->assertTrue($reportListSettings->isZebraTable());
     }
 }
